chore(carousel): renaming enums file
