+++ projects/igniteui-angular/src/lib/carousel/enums.ts
